Initial revision of leap in Go
First off, I'm not a Gopher! I only recently discovered the language
so if anybody has any advice which would make my Go more idiomatic,
I would really appreciate it.

I'm in the camp of limiting the number of return statements in a
method. I'm not against using more than one return statement but
usually prefer the result variable style.

I was mostly struggling with how to structure the if statement. I'm
not really happy with it as it is as I think it's too concise for
its own good but I don't like nested if statements either. Still
pondering what I would prefer...

See the exercism.io submission for context:
http://exercism.io/submissions/45221bb13c20254a9d250618