Second revision of leap in Go
The previous revision was too smart for its own good. It was concise but
readability suffered.

I let go of the single return statement rule and it improved the code in
this instance. I think it works for a problem of this scale but if the
function would have been longer or more complex it would have obfuscated
the flow of the function.

See the exercism.io submission for context:
http://exercism.io/submissions/16fcec55c7a6a45c9fd6d926